requirements.txt
yfinance
pandas
gspread
google-auth